Harden CI checks
We end up committing a whole bunch of warnings all the time, which
create unnecessary noise for others. This turns them into errors, to
make sure that none of those slip through the cracks.

Also, any segmentation faults or aborts didn't cause the tests to fail
in the past, which should hopefully be fixed now.