Snapshot heap-backed tables for the prelude cache.
Previously we couldn't have table data in the prelude. Tables are represented as
lamda terms that include heap memory pointers. These pointer values are
meaningless in a new process when we deserialize them from the on-disk cache.

This change traverses pointer literals in to-be-cached data and loads
serializable byte strings from the memory they point to (either directly or
indirectly). This means we need to track the sizes of allocations. We do that by
making `malloc_dex` store the allocation size in a 64-byte metadata prefix
before the payload part of the allocation.