Add an effect narrowing pass
At the moment we are very imprecise in the effect rows we use to
annotate core lambda expressions and usually just stick the full set of
allowed effects in there without any regard for whether the effects are
actually caused by the body. The new pass narrows those annotations as
much as possible, which can be useful for both inlining and
parallelization decisions.